- Share SNS publisher between all topic publishers that share a topic name customiser
