+++ src/JustSaying/Fluent/PublishConfig/DynamicMessagePublisher.cs
@@ -1,0 +1,71 @@
+using System.Collections.Concurrent;
+using JustSaying.Messaging;
+using JustSaying.Messaging.Interrogation;
+using JustSaying.Models;
+using Microsoft.Extensions.Logging;
+
+namespace JustSaying.Fluent;
+
+internal class DynamicMessagePublisher : IMessagePublisher
+{
+    private readonly ConcurrentDictionary<string, IMessagePublisher> _publisherCache = new();
+    private readonly Func<Message, string> _topicNameCustomizer;
+    private readonly Func<string, StaticPublicationConfiguration> _staticConfigBuilder;
+
+    private readonly ConcurrentDictionary<string, SemaphoreSlim> _topicCreationLocks = new();
+    private readonly ILogger<DynamicMessagePublisher> _logger;
+
+    public DynamicMessagePublisher(Func<Message, string> topicNameCustomizer, Func<string, StaticPublicationConfiguration> staticConfigBuilder, ILoggerFactory loggerFactory)
+    {
+        _topicNameCustomizer = topicNameCustomizer;
+        _staticConfigBuilder = staticConfigBuilder;
+        _logger = loggerFactory.CreateLogger<DynamicMessagePublisher>();
+    }
+
+    public InterrogationResult Interrogate()
+    {
+        return new InterrogationResult(new
+        {
+
+        });
+    }
+
+    public Task StartAsync(CancellationToken stoppingToken)
+    {
+        return Task.CompletedTask;
+    }
+
+    public async Task PublishAsync(Message message, PublishMetadata metadata, CancellationToken cancellationToken)
+    {
+        var topicName = _topicNameCustomizer(message);
+        if (_publisherCache.TryGetValue(topicName, out var publisher))
+        {
+            await publisher.PublishAsync(message, metadata, cancellationToken).ConfigureAwait(false);
+            return;
+        }
+
+        var lockObj = _topicCreationLocks.GetOrAdd(topicName, _ => new SemaphoreSlim(1, 1));
+
+        _logger.LogDebug("Publisher for topic {TopicName} not found, waiting on creation lock", topicName);
+        await lockObj.WaitAsync(cancellationToken).ConfigureAwait(false);
+        if (_publisherCache.TryGetValue(topicName, out var thePublisher))
+        {
+            _logger.LogDebug("Lock re-entrancy detected, returning existing publisher");
+            await thePublisher.PublishAsync(message, metadata, cancellationToken).ConfigureAwait(false);
+            return;
+        }
+
+        _logger.LogDebug("Lock acquired to init topic {TopicName}", topicName);
+        var config = _staticConfigBuilder(topicName);
+        _logger.LogDebug("Executing startup task for topic {Top ame}", topicName);
+        await config.StartupTask(cancellationToken).ConfigureAwait(false);
+
+        _publisherCache.TryAdd(topicName, config.Publisher);
+
+        _logger.LogDebug("Publishing message on newly created topic {TopicName}", topicName);
+        await config.Publisher.PublishAsync(message, metadata, cancellationToken).ConfigureAwait(false);
+    }
+
+    public Task PublishAsync(Message message, CancellationToken cancellationToken)
+        => PublishAsync(message, null, cancellationToken);
+}
